refactore2
