+++ Lesson6_2/app/src/main/java/com/artem/lesson6_2/data/Note.java
@@ -1,4 +1,4 @@
-package com.artem.lesson6_2;
+package com.artem.lesson6_2.data;
 
 import android.widget.ImageView;
