+++ Lesson6_2/app/src/main/java/com/artem/lesson6_2/pages/DatePickerFragment.java
@@ -1,4 +1,4 @@
-package com.artem.lesson6_2;
+package com.artem.lesson6_2.pages;
 
 import android.app.DatePickerDialog;
 import android.app.Dialog;
