+++ Lesson6_2/app/src/main/res/layout/fragment_about.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AboutFragment">
+    tools:context=".pages.AboutFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
