+++ Lesson6_2/app/src/main/res/layout/fragment_add_note.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".EditFragment">
+    tools:context=".pages.EditFragment">
 
 
     <LinearLayout
