+++ Lesson6_2/app/src/main/res/layout/fragment_date_picker.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DatePickerFragment">
+    tools:context=".pages.DatePickerFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
