+++ Lesson6_2/app/src/main/res/layout/fragment_done_notes.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:orientation="vertical"
     android:layout_height="match_parent"
-    tools:context=".DoneNotesFragment">
+    tools:context=".pages.DoneNotesFragment">
 
     <TextView
         android:id="@+id/text_no_done"
