+++ Lesson6_2/app/src/main/res/layout/fragment_edit.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".EditFragment">
+    tools:context=".pages.EditFragment">
 
 
     <LinearLayout
